+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Revision history for web-view
 
+## 0.7.0
+
+* stack, popup, offset, layer - more intuitive interface
+* added Web.View.Url.renderPath
+* Style class
+* added code, lists
+
 ## 0.6.0
 
 * stack - layout children on top of each other